backend/requirements.txt
Faker==25.5.0
Flask==3.0.3
Flask_Cors==4.0.1
Flask_SocketIO==5.3.6
Pillow==10.3.0
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
